+++ cached_network_image/lib/src/image_provider/cached_network_image_provider.dart
@@ -67,59 +67,7 @@ class CachedNetworkImageProvider
     return SynchronousFuture<CachedNetworkImageProvider>(this);
   }
 
-  @Deprecated(
-    'load is deprecated, use loadImage instead, '
-    'see https://docs.flutter.dev/release/breaking-changes/image-provider-load-buffer',
-  )
-  @override
-  ImageStreamCompleter load(
-    CachedNetworkImageProvider key,
-    DecoderCallback decode,
-  ) {
-    final chunkEvents = StreamController<ImageChunkEvent>();
-    return MultiImageStreamCompleter(
-      codec: _loadAsync(key, chunkEvents, decode),
-      chunkEvents: chunkEvents.stream,
-      scale: key.scale,
-      informationCollector: () sync* {
-        yield DiagnosticsProperty<ImageProvider>(
-          'Image provider: $this \n Image key: $key',
-          this,
-          style: DiagnosticsTreeStyle.errorProperty,
-        );
-      },
-    );
-  }
-
-  @Deprecated(
-    '_loadAsync is deprecated, use loadImage instead, '
-    'see https://docs.flutter.dev/release/breaking-changes/image-provider-load-buffer',
-  )
-  Stream<ui.Codec> _loadAsync(
-    CachedNetworkImageProvider key,
-    StreamController<ImageChunkEvent> chunkEvents,
-    DecoderCallback decode,
-  ) {
-    assert(key == this);
-    return ImageLoader().loadAsync(
-      url,
-      cacheKey,
-      chunkEvents,
-      decode,
-      cacheManager ?? DefaultCacheManager(),
-      maxHeight,
-      maxWidth,
-      headers,
-      errorListener,
-      imageRenderMethodForWeb,
-      () => PaintingBinding.instance.imageCache.evict(key),
-    );
-  }
-
-  @Deprecated(
-    'loadBuffer is deprecated, use loadImage instead, '
-    'see https://docs.flutter.dev/release/breaking-changes/image-provider-load-buffer',
-  )
+  @Deprecated('loadBuffer is deprecated, use loadImage instead')
   @override
   ImageStreamCompleter loadBuffer(
     CachedNetworkImageProvider key,