+++ cached_network_image_platform_interface/CHANGELOG.md
@@ -1,7 +1,9 @@
 ## [2.0.1] - TBD
-* Support for Dart 3
-* Use loadImage instead loadBuffer and load
-* Use ImageDecoderCallback with ImageProvider.loadImage instead of DecoderBufferCallback
+* Specify types
+* Add error to ErrorListener
+* Update example
+* Use `loadImage` instead `loadBuffer`
+* Remove [`load`](https://github.com/flutter/flutter/pull/132679)
 
 ## [2.0.0] - 2022-08-31
 * Added loadBufferAsync for Flutter 3.3