+++ cached_network_image_web/CHANGELOG.md
@@ -1,7 +1,9 @@
 ## [1.0.3] - TBD
-* Support for Dart 3
-* Use loadImage instead loadBuffer and load
-* Use ImageDecoderCallback with ImageProvider.loadImage instead of DecoderBufferCallback
+* Specify types
+* Add error to ErrorListener
+* Update example
+* Use `loadImage` instead `loadBuffer`
+* Remove [`load`](https://github.com/flutter/flutter/pull/132679)
 
 ## [1.0.2] - 2022-08-31
 * Added loadBufferAsync and deprecated loadAsync